core/core/urls.py
from django.urls import path, include
from rest_framework import routers
from django.contrib import admin
from android import views
from android.views import SignupView
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenBlacklistView,
)


urlpatterns = [
    path('admin/', admin.site.urls),
    path('login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('logout/', TokenBlacklistView.as_view(), name='token_blacklist'),
    path('signup/', SignupView.as_view(), name='signup'),
    path('api/', include('android.urls')),
]
