core/requirements.txt
django
djangorestframework
django-cors-headers
pillow
django-filter
djoser 
djangorestframework-simplejwt